[sw/device/tests] Prepare to add C warning flags

+++ sw/device/tests/pwrmgr_wdog_reset_reqs_test.c
@@ -52,15 +52,15 @@ static void config_wdog(const dif_aon_timer_t *aon_timer,
  * Execute the aon timer wdog bite reset test.
  */
 static void wdog_bite_test(const dif_aon_timer_t *aon_timer,
-                           const dif_pwrmgr_t *pwrmgr, uint64_t bark_time_us) {
+                           const dif_pwrmgr_t *pwrmgr, uint32_t bark_time_us) {
   uint64_t bite_time_us = bark_time_us * 2;
   config_wdog(aon_timer, pwrmgr, bark_time_us, bite_time_us);
 
   // The `intr_state` takes 3 aon clock cycles to rise plus 2 extra cycles as a
   // precaution.
   uint32_t wait_us =
-      bark_time_us +
-      udiv64_slow(5 * 1000000 + kClockFreqAonHz - 1, kClockFreqAonHz, NULL);
+      bark_time_us + (uint32_t)udiv64_slow(5 * 1000000 + kClockFreqAonHz - 1,
+                                           kClockFreqAonHz, NULL);
 
   // Wait bark time and check that the bark interrupt requested.
   busy_spin_micros(wait_us);
